requirements.txt
beautifulsoup4
imap-tools
openai
python-dotenv
